Create OrangeBtn and GrayBtn files to split code in diferent files Optional
